ingest/requirements.txt
appnope
asttokens
backcall
cachetools
certifi
charset-normalizer
click
comm
debugpy
decorator
executing
Flask
google-api-core
google-auth
google-cloud-bigquery
google-cloud-core
google-cloud-storage
google-cloud-logging
google-crc32c
google-resumable-media
googleapis-common-protos
grpcio
grpcio-status
gunicorn
idna
ipykernel
ipython
itsdangerous
jedi
Jinja2
jupyter_client
jupyter_core
MarkupSafe
matplotlib-inline
nest-asyncio
numpy
packaging
pandas
parso
pathlib
pexpect
pickleshare
pip
platformdirs
prompt-toolkit
proto-plus
protobuf
psutil
ptyprocess
pure-eval
pyarrow
pyasn1
pyasn1-modules
Pygments
python-dateutil
pytz
PyYAML
pyzmq
requests
rsa
setuptools
six
stack-data
tornado
traitlets
tzdata
urllib3
wcwidth
Werkzeug
wheel
xmltodict
structlog
pywinauto
pysftp
pathlib
python-dotenv
fsspec
gcsfs